Add arena logic on main scene
