+++ src/shared/lib/ToggleableList/index.ts
@@ -1,0 +1,2 @@
+export type { ToggleableItemI as ToggleableItem, ToggleableItemRawI as ToggleableItemRaw } from './ToggleableList'
+export { ToggleableList } from './ToggleableList'
